Some pre-integration refactoring of utbot-spring-analyzer
