+++ utbot-spring-analyzer/src/main/kotlin/org/utbot/spring/config/TestApplicationConfiguration.kt
@@ -1,9 +1,9 @@
-package config
+package org.utbot.spring.config
 
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import post_processors.UtBotBeanFactoryPostProcessor
+import org.utbot.spring.post_processors.UtBotBeanFactoryPostProcessor
 
 @Configuration
 open class TestApplicationConfiguration {
