+++ utbot-spring-analyzer/src/main/kotlin/org/utbot/spring/configurators/XmlFilesConfigurator.kt
@@ -1,9 +1,7 @@
-package application.configurators
+package org.utbot.spring.configurators
 
-import analyzers.XmlConfigurationAnalyzer
-import application.utils.FakeFileManager
-import utils.ConfigurationManager
-import utils.PathsUtils
+import org.utbot.spring.utils.ConfigurationManager
+import org.utbot.spring.utils.PathsUtils
 import kotlin.io.path.Path
 
 class XmlFilesConfigurator(
@@ -17,10 +15,10 @@ class XmlFilesConfigurator(
         for (userXmlFilePath in userXmlFilePaths) {
             if(userXmlFilePath == PathsUtils.EMPTY_PATH)continue
 
-            val xmlConfigurationAnalyzer =
-                XmlConfigurationAnalyzer(userXmlFilePath, PathsUtils.createFakeFilePath(userXmlFilePath))
+            val xmlConfigurationParser =
+                XmlConfigurationParser(userXmlFilePath, PathsUtils.createFakeFilePath(userXmlFilePath))
 
-            xmlConfigurationAnalyzer.fillFakeApplicationXml()
+            xmlConfigurationParser.fillFakeApplicationXml()
             configurationManager.patchImportResourceAnnotation(Path(userXmlFilePath).fileName)
         }
     }
