+++ utbot-spring-analyzer/src/main/kotlin/org/utbot/spring/utils/PathsUtils.kt
@@ -1,5 +1,6 @@
-package utils
+package org.utbot.spring.utils
 
+import java.io.File
 import kotlin.io.path.Path
 
 object PathsUtils {
@@ -8,14 +9,13 @@ object PathsUtils {
     fun createFakeFilePath(fileName: String): String =
         Path(buildResourcesPath, "fake_${Path(fileName).fileName}").toString()
 
-    //TODO: it is better to do it without marker files
     private val buildResourcesPath: String
         get() {
             val resourcesMarker =
                 this.javaClass.classLoader.getResource("resources_marker.txt")
                     ?: error("Resources marker file is not found")
 
-            return Path(resourcesMarker.path).parent.toString()
+            return Path(File(resourcesMarker.toURI()).path).parent.toString()
         }
 
 }
