+++ utbot-spring-analyzer/src/main/resources/resources_marker.txt
